requirements.txt
python-dotenv==1.0.1
telebot==0.0.5
yt-dlp==2024.12.13